Oprava chyb umožňující ovlivňování cizích obrázků skrz správu hlášení
Při manuální úpravě argumentů předávaných JavaScriptové funkci (nebo odesíláním AJAX požadavků s ručně nastaveným ID obrázku nebo hlášení) mohlo být smazáno hlášení nebo obrázek, nebo mohl být obrázek skryt či upraven, i když daný subjekt nespadal do právě spravované třídy.
Nyní se již vždy kontroluje, zda je daný obrázek součástí nějaké poznávačky ve spravované třídě, stejně tak jestli se k takovému obrázku vztahuje odstraňované hlášení.

+++ Models/ReportResolver.php
@@ -5,11 +5,12 @@
  */
 class ReportResolver
 {
+    private $class;
     private $group;
     
     /**
      * Konstruktor zajišťující, že instanci této třídy lze vytvořit pouze pokud je přihlášen správce zvolené třídy
-     * Také je nastavena poznávačka, ve které je možné vytvořenou instancí řešit hlášení
+     * Také je nastavena poznávačka, ve které je možné vytvořenou instancí řešit hlášení a třída, do které musí spadat přírodniny, k jejímž obrázkům se všechna hlášení vztahují
      * @throws AccessDeniedException V případě, že není vybrána žádná třída nebo skupina nebo pokud přihlášený uživatel není správcem zvolené třídy
      */
     public function __construct()
@@ -30,6 +31,7 @@ class ReportResolver
         {
             throw new AccessDeniedException(AccessDeniedException::REASON_GROUP_NOT_CHOSEN);
         }
+        $this->class = $class;
         $this->group = $_SESSION['selection']['group'];
     }
     
@@ -44,8 +46,15 @@ class ReportResolver
     public function editPicture(int $pictureId, string $newNaturalName, string $newUrl): void
     {
         $picture = new Picture(false, $pictureId);
+        
+        //Kontrola, zda je vypínaný obrázek součástí nějaké přírodniny patřící do spravované třídy
+        if ($this->checkPictureBelongsToClass($picture))
+        {
+            throw new AccessDeniedException(AccessDeniedException::REASON_MANAGEMENT_REPORTS_RESOLVE_PICTURE_FOREIGN_NATURAL);
+        }
+        
         $natural = new Natural(false, 0);
-        $natural->initialize($newNaturalName, null, null, $this->group->getClass());
+        $natural->initialize($newNaturalName, null, null, $this->class);
         
         try
         {
@@ -75,6 +84,11 @@ class ReportResolver
     public function disablePicture(int $pictureId): void
     {
         $picture = new Picture(false, $pictureId);
+        //Kontrola, zda je vypínaný obrázek součástí nějaké přírodniny patřící do spravované třídy
+        if ($this->checkPictureBelongsToClass($picture))
+        {
+            throw new AccessDeniedException(AccessDeniedException::REASON_MANAGEMENT_REPORTS_RESOLVE_PICTURE_FOREIGN_NATURAL);
+        }
         $picture->disable();
         $picture->deleteReports();
     }
@@ -86,6 +100,11 @@ class ReportResolver
     public function deletePicture(int $pictureId): void
     {
         $picture = new Picture(false, $pictureId);
+        //Kontrola, zda je vypínaný obrázek součástí nějaké přírodniny patřící do spravované třídy
+        if ($this->checkPictureBelongsToClass($picture));
+        {
+            throw new AccessDeniedException(AccessDeniedException::REASON_MANAGEMENT_REPORTS_RESOLVE_PICTURE_FOREIGN_NATURAL);
+        }
         $picture->delete();
     }
     
@@ -96,6 +115,21 @@ class ReportResolver
     public function deleteReport(int $reportId): void
     {
         $report = new Report(false, $reportId);
+        //Kontrola, zda se odstraňované hlášení vztahuje k obrázku, který je součástí nějaké přírodniny patřící do spravované třídy
+        if (!$this->checkPictureBelongsToClass($report->getPicture()));
+        {
+            throw new AccessDeniedException(AccessDeniedException::REASON_MANAGEMENT_REPORTS_RESOLVE_PICTURE_FOREIGN_NATURAL);
+        }
         $report->delete();
+    }
+    
+    /**
+     * Metoda kontrolující, zda je daný obrázek přiřazen k přírodnině, která je součástí nějaké poznávačky spravované třídy 
+     * @param Picture $picture Obrázek pro kontrolu
+     * @return bool TRUE, pokud je obrázek součástí nějaké poznávačky patřící do spravované třídy, FALSE, pokud ne
+     */
+    private function checkPictureBelongsToClass(Picture $picture): bool
+    {
+        return ($picture->getNatural()->getClass()->getId() === $this->class->getId());
     }
 }
